backend/requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
flask-cors>=4.0.0
PyJWT>=2.8.0
sentence-transformers>=2.6.1
joblib>=1.3.2
scikit-learn>=1.3.2
gunicorn>=21.2.0
psycopg2-binary>=2.9.9 